+++ doc/type-search-index.js
@@ -1,1 +1,0 @@
-typeSearchIndex = [{"l":"All Classes and Interfaces","u":"allclasses-index.html"},{"p":"org.bluezoo.gonzalez","l":"NamespaceScopeTracker"},{"p":"org.bluezoo.gonzalez","l":"Parser"}];updateSearchResults();